+++ lib/wanikani.rb
@@ -8,6 +8,7 @@ require 'wanikani/level'
 require 'wanikani/srs'
 require 'wanikani/recent_unlocks'
 require 'wanikani/critical_items'
+require 'wanikani/exceptions'
 
 Encoding.default_external = Encoding::UTF_8
 Encoding.default_internal = Encoding::UTF_8
@@ -52,6 +53,6 @@ module Wanikani
   private
 
   def self.raise_exception(message)
-    raise Exception, "There was an error fetching the data from Wanikani (#{message})"
+    raise Wanikani::Exception, "There was an error fetching the data from Wanikani (#{message})"
   end
 end