+++ lib/wanikani/exceptions.rb
@@ -1,0 +1,3 @@
+module Wanikani
+  class Exception < Exception; end
+end